Create hook to keep code DRY
